Flatten Settings sidebar; unify Engine & Model pane
- Drop sidebar section headers and the Apple Intelligence / Open Source
  sub-rows. The flat list is shorter and tighter, and removing the
  disclosure-chevron gutter lets the sidebar shrink to ~200pt without
  truncating labels.
- Collapse the two engine sub-panes into one Engine & Model pane. The
  engine dropdown doubles as the in-pane switcher: pick a row, see its
  settings. Removes the duplicated "Switch to..." buttons.
- Cap every detail pane's content at a centered 600pt column inside
  SettingsPaneScaffold so toggles sit next to their labels instead of
  stretching across the window.
- Migrate persisted sidebar selection: users on the old appleIntelligence
  or openSource raw values land on the unified Engine & Model pane.
- Consolidate SettingsAttentionEvaluator onto .engineAndModel; update
  tests to match.
