+++ Cotabby/UI/Settings/Components/SettingsPaneScaffold.swift
@@ -10,6 +10,13 @@ import SwiftUI
 /// surfaces attention per pane: when a pane is in a degraded state (missing permission, runtime
 /// unavailable) we render an inline callout above the form so the actionable surface lives next to
 /// the controls that fix it.
+/// Centers the pane's grouped form in a fixed-width column inside the detail pane. macOS's own
+/// System Settings uses the same trick: rows sit in a comfortable reading column rather than
+/// stretching to the full window width, so toggles don't end up far away from their labels when
+/// the user resizes the window. Lives at file scope because Swift forbids static stored
+/// properties on generic types.
+private let settingsPaneContentMaxWidth: CGFloat = 600
+
 struct SettingsPaneScaffold<Content: View>: View {
     let callout: SettingsPaneCallout?
     @ViewBuilder let content: () -> Content
@@ -40,6 +47,8 @@ struct SettingsPaneScaffold<Content: View>: View {
                 // breathing room regardless of whether the first section carries a header.
                 .padding(.top, 12)
             }
+            .frame(maxWidth: settingsPaneContentMaxWidth)
+            .frame(maxWidth: .infinity)
         }
     }
 }
