+++ CotabbyTests/SettingsAttentionEvaluatorTests.swift
@@ -33,9 +33,9 @@ final class SettingsAttentionEvaluatorTests: XCTestCase {
         XCTAssertEqual(categories, [.permissions])
     }
 
-    /// Apple Intelligence unavailability flags both the sub-row and its parent so the user sees
-    /// the dot regardless of which row their eye lands on first.
-    func test_appleIntelligenceUnavailable_flagsSubRowAndParent() {
+    /// Apple Intelligence unavailability flags the unified Engine & Model row — the dedicated
+    /// sub-row was removed when the sidebar was flattened.
+    func test_appleIntelligenceUnavailable_flagsEngineAndModel() {
         let categories = SettingsAttentionEvaluator.categoriesNeedingAttention(
             makeInputs(
                 selectedEngine: .appleIntelligence,
@@ -43,29 +43,28 @@ final class SettingsAttentionEvaluatorTests: XCTestCase {
                 foundationModelMessage: "Apple Intelligence is turned off in System Settings."
             )
         )
-        XCTAssertEqual(categories, [.appleIntelligence, .engineAndModel])
+        XCTAssertEqual(categories, [.engineAndModel])
     }
 
     /// The flag is engine-scoped: if the user is on Open Source, FM availability doesn't matter.
-    func test_appleIntelligenceUnavailable_butLlamaSelected_noAppleAttention() {
+    func test_appleIntelligenceUnavailable_butLlamaSelected_noEngineAttention() {
         let categories = SettingsAttentionEvaluator.categoriesNeedingAttention(
             makeInputs(
                 selectedEngine: .llamaOpenSource,
                 foundationModelAvailable: false
             )
         )
-        XCTAssertFalse(categories.contains(.appleIntelligence))
         XCTAssertFalse(categories.contains(.engineAndModel))
     }
 
-    func test_llamaRuntimeFailed_flagsOpenSourceAndParent() {
+    func test_llamaRuntimeFailed_flagsEngineAndModel() {
         let categories = SettingsAttentionEvaluator.categoriesNeedingAttention(
             makeInputs(
                 selectedEngine: .llamaOpenSource,
                 llamaRuntimeFailedReason: "Model failed to load."
             )
         )
-        XCTAssertEqual(categories, [.openSource, .engineAndModel])
+        XCTAssertEqual(categories, [.engineAndModel])
     }
 
     func test_callout_permissions_returnsActionableMessage() {
@@ -77,23 +76,33 @@ final class SettingsAttentionEvaluatorTests: XCTestCase {
         XCTAssertTrue(message?.contains("more access") ?? false)
     }
 
-    func test_callout_appleIntelligence_echoesAvailabilityMessage() {
+    func test_callout_engineAndModel_appleIntelligence_echoesAvailabilityMessage() {
         let inputs = makeInputs(
             selectedEngine: .appleIntelligence,
             foundationModelAvailable: false,
             foundationModelMessage: "This Mac is not eligible for Apple Intelligence."
         )
-        let message = SettingsAttentionEvaluator.calloutMessage(for: .appleIntelligence, inputs: inputs)
+        let message = SettingsAttentionEvaluator.calloutMessage(for: .engineAndModel, inputs: inputs)
         XCTAssertEqual(message, "This Mac is not eligible for Apple Intelligence.")
     }
 
-    func test_callout_openSource_echoesRuntimeFailureReason() {
+    func test_callout_engineAndModel_openSource_echoesRuntimeFailureReason() {
         let inputs = makeInputs(
             selectedEngine: .llamaOpenSource,
             llamaRuntimeFailedReason: "Couldn't open the GGUF file."
         )
-        let message = SettingsAttentionEvaluator.calloutMessage(for: .openSource, inputs: inputs)
+        let message = SettingsAttentionEvaluator.calloutMessage(for: .engineAndModel, inputs: inputs)
         XCTAssertEqual(message, "Couldn't open the GGUF file.")
+    }
+
+    func test_callout_engineAndModel_healthyEngine_isNil() {
+        let inputs = makeInputs(
+            selectedEngine: .appleIntelligence,
+            foundationModelAvailable: true
+        )
+        XCTAssertNil(
+            SettingsAttentionEvaluator.calloutMessage(for: .engineAndModel, inputs: inputs)
+        )
     }
 
     func test_callout_paneWithoutAttention_isNil() {
@@ -104,17 +113,5 @@ final class SettingsAttentionEvaluatorTests: XCTestCase {
                 "\(category) should never carry a callout"
             )
         }
-    }
-
-    func test_callout_engineAndModel_neverReturnsItsOwnMessage() {
-        // The parent pane gets a sidebar dot but the actionable message lives on the sub-pane.
-        let inputs = makeInputs(
-            selectedEngine: .appleIntelligence,
-            foundationModelAvailable: false,
-            foundationModelMessage: "Apple Intelligence is turned off."
-        )
-        XCTAssertNil(
-            SettingsAttentionEvaluator.calloutMessage(for: .engineAndModel, inputs: inputs)
-        )
     }
 }
